フィードバック内容の修正と在庫属性名の変更

+++ web/database/migrations/2020_04_22_143136_carts.php
@@ -16,8 +16,8 @@ class Carts extends Migration
         //
         Schema::create('carts', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('product_id');
             $table->unsignedInteger('quantity');
             $table->timestamps();
